ENH: Also keep whitespace in comma-separated arguments and input stream
For the input steam, enhance the sed program to differentiate between comma and whitespace separators; this avoids that whitespace also gets eaten when comma separation is used. (Unfortunately, we cannot return comma-separated newlines due to readarray swallowing those; however, this would also clash with the use case of breaking up long lines into multiple ones (only treat ,\n, as newline, but not \n, or ,\n), and multiple newlines would clash with the stated rule that completely empty lines resolve to a space.

For the argument(s), replace the array definition (bad style in any case; "set noglob" was missing to make that more robust) with a read call. As that one only handles a single line, replace newlines with a temporary ^G placeholder, split by command into an array, and then add the array elements with the placeholder converted back to newline.