ENH: Minor: Allow specifying default ticks --as default or --as bars
And use those even when the ticks are overridden via $SPARK_TICKS. A command-line parameter for the default tick is useful when an application chooses a different tick, but allows users to override that; then, --as bars gives back the original style.
P.S.: This also allows coloring (--as green-bars), but that combination doesn't look very useful.