CHG: Change default division algorithm to even distribution, add --single-max-range to switch back
While looking into how to implement a legend of values, I accidentally found out that the distribution of values into buckets only groups the highest value into the top bucket! I had assumed an even distribution. Dunno if this was done intentionally or is a bug?!
The algorithm is easy to fix; add a evenDivider() function and make that the default.

Tests: Split off separate tests that highlight the differences. Allow invoking tests directly from the project dir.