Refactoring: Move into bin, replace roundup with Bats, reuse scripting-setup workflow
Roundup hasn't seen any development in 10 years, and doesn't offer any benefits over the more common Bats.
Switching to my scripting-setup further reduces the boilerplate code. The setup requires the command to be in ./bin, which already was documented that way, but not the case.

System-info: Ubuntu 20.04.6 LTS, x86_64
Platform-info: bash 5.0.17(1)-release, Bats 1.10.0